requirements.txt
playwright>=1.41.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.24.0
supabase>=2.4.0
python-dotenv>=1.0.0
